requirements.txt
# Core OpenVINO and Optimization
openvino-dev>=2024.0.0
optimum-intel[openvino,nncf]>=1.15.0

# LLM interaction
transformers>=4.38.0
sentencepiece
readline      # for nice terminal input
